requirements.txt
ipykernel
langchain
python-dotenv
langchain-openai
pylance
beautifulsoup4
langchain_community
faiss-cpu